openshift: Prepare for database-restore deployment
We are going for one Dockerfile and build context for two deployments.  First
deployment "database" will, as before, have the live database running.  The
other deployment "database-restore" will run pg_restore to load the next
database generation.  Which deployment is build will be controlled by the
DATABASE_RESTORE build configuration env variable.  If DATABASE_RESTORE is set
to a non-empty value we build for the "database-restore" deployment, otherwise
for the "database" deployment.  For the build the difference is only in the
postgres configuration.  For the running pods there is a difference of
code-path in the new start-database script.

Before running or re-starting any pods the PGDATA_GENERATION environment
variable needs to be set appropriately.

Changes to database:
  * local.conf has settings for both containers
  * restore_mode.conf has settings for running pg_restore
  * $PGDATA is determined by PGDATA_GENERATION container runtime variable
  * start-database script added (can initialize new databases for
    restore mode, does some sanity checking)
  * stop-database script for graceful shutdown